+++ src/Vouch.php
@@ -531,6 +531,8 @@ class Vouch extends Plugin
                 $event->rules['POST vouch/sources/test'] = 'vouch/sources/test';
                 $event->rules['POST vouch/sources/sync'] = 'vouch/sources/sync';
                 $event->rules['POST vouch/sources/find-google-place'] = 'vouch/sources/find-google-place';
+                $event->rules['vouch/sources/connect-google'] = 'vouch/sources/connect-google';
+                $event->rules['vouch/sources/google-oauth-callback'] = 'vouch/sources/google-oauth-callback';
 
                 $event->rules['vouch/settings'] = 'vouch/settings/edit';
                 $event->rules['POST vouch/settings'] = 'vouch/settings/save';